removed flask, updated deploy
