--- flask_server/README.md
@@ -1,7 +1,0 @@
-# Серверная часть
-
-Отвечает за взаимодействие с БД и регистрацию/авторизацию пользователей для игры.
-
-Общий вид представлен ниже:
-
-<img src='./server_card.png'>
