--- flask_server/app.py
@@ -1,15 +1,0 @@
-import logging
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-
-from config import SQLALCHEMY_DATABASE_URI, SQLALCHEMY_TRACK_MODIFICATIONS
-
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = SQLALCHEMY_DATABASE_URI
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = SQLALCHEMY_TRACK_MODIFICATIONS
-app.logger.addHandler(logging.StreamHandler())
-app.logger.setLevel(logging.INFO)
-
-db = SQLAlchemy(app)
-
-from models import *  # NOQA
