--- flask_server/requirements.txt
@@ -1,7 +1,0 @@
-Flask==0.10.1
-Flask-SQLAlchemy==2.1
-gunicorn==19.6.0
-SQLAlchemy==1.0.13
-psycopg2==2.6.2
-coverage==4.2
-requests==2.11.1
