--- flask_server/server.py
@@ -1,5 +1,0 @@
-from app import app
-import views  # NOQA
-
-if __name__ == '__main__':
-    app.run(port=8000, debug=True)
